+++ src/agentx/tool_registry.py
@@ -99,6 +99,7 @@ enabled = true
 reload_tools = true
 register_tool = true
 diagnose_tools = true
+open_file_in_editor = true
 """